+++ internal/client/ovh_clusters.go
@@ -46,6 +46,7 @@ type CreateOvhClusterRequest struct {
 	EtcdFlavorID          string  `json:"etcd_flavor_id,omitempty"`
 	ExternalCloudProvider bool    `json:"external_cloud_provider"`
 	IncludeNetworking     bool    `json:"include_networking"`
+	IncludeDNS            bool    `json:"include_dns"`
 	GitopsCredentialName  *string `json:"gitops_credential_name,omitempty"`
 	GitopsRepository      *string `json:"gitops_repository,omitempty"`
 	GitopsBranch          *string `json:"gitops_branch,omitempty"`